feat: Added a hook for move piece for better coding conventions and added comments.
