+++ frontend/src/features/board/hooks/useRecordMove.ts
@@ -1,3 +1,4 @@
+// ---------------- Imports ---------------- //
 import { useMoveLog } from '../../history/moveLogStore';
 
 export function useRecordMove() {
